version estabilizada, iniciando refactor de escenarios

+++ include/general/basic_entities.h
@@ -10,9 +10,9 @@ class e_base
 public:
     basic_state entity_state;
 
-    float mass;
-    float size_r;
-    float selectable_size = 25;
+    double mass;
+    double size_r;
+    double selectable_size = 25;
 
     f_vector main_dimensions;
     f_vector inertia_tensor;
@@ -22,8 +22,8 @@ protected:
 
 public:
     e_base() : mass(0), size_r(0), main_dimensions(3), inertia_tensor(9) {};
-    e_base(float mass, float size_r, basic_state entity_state) : mass(mass), size_r(size_r), entity_state(entity_state) {};
-    e_base(float mass, float size_r, basic_state entity_state, f_vector main_dimensions, f_vector inertia_tensor) : mass(mass), size_r(size_r), entity_state(entity_state), main_dimensions(main_dimensions), inertia_tensor(inertia_tensor) {};
+    e_base(double mass, double size_r, basic_state entity_state) : mass(mass), size_r(size_r), entity_state(entity_state) {};
+    e_base(double mass, double size_r, basic_state entity_state, f_vector main_dimensions, f_vector inertia_tensor) : mass(mass), size_r(size_r), entity_state(entity_state), main_dimensions(main_dimensions), inertia_tensor(inertia_tensor) {};
 
     e_base(const e_base &rh) : mass(rh.mass), size_r(rh.size_r), entity_state(rh.entity_state), main_dimensions(rh.main_dimensions), inertia_tensor(rh.inertia_tensor) {};
     e_base(e_base &&rh) : mass(rh.mass), size_r(rh.size_r), entity_state(std::move(rh.entity_state)), main_dimensions(std::move(rh.main_dimensions)), inertia_tensor(std::move(rh.inertia_tensor)) {};
